Move map permissions handling to HomeFragment
