+++ app/src/main/java/com/qrcode_quest/ui/map/MapFragment.java
@@ -248,7 +248,6 @@ public class MapFragment extends Fragment {
     }
 
 
-
     @Override
     public void onResume(){
         super.onResume();
